Added the ability to register tasks and ability to context switch
Notes on this:
Because we don't have a heap allocator everything is staticaly defined
There should be a max of ten tasks, but because I needed a place to store
the regs from the initial process during the first context switch, I for
some reason chose to do that in the entry for task 7, there are only a max
of 7 tasks.
The scheduler is round robin.

What needs to be done:
    * Add a more elegant way to do the first context switch
    * Add a way allow tasks to sleep
    * Add a heap allocator
    * Allow for an arbitrary amount of tasks as memory allows

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>