Have TASK[0] store the information of the initial thread of execution
This is needed because the context switch function always pushes
registers R4-R11 to the stack and saves the stack pointer of the
currently running process, so we have to store the stuff from the initial
thread here. This feels a little sloppy, so maybe we should add logic
to avoid saving the context of the very first process.

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>