Fix bug where heap block header and footers would be overwritten
Before when a block was being allocated from a free block, it was always
assumed that the block being allocated from was larger than the size requested,
so the allocator would always try to "split" the block into the allocated and
unallocated blocks. This was problematic when the block we are allocating from
was the same size as what was requested, causing data to be overwritten.
This problem is fixed with this commit.

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>