Unmark tasks as queued if the scheduler queue is empty
For whatever reason, sometimes tasks would be marked as queued even though
they were not actually in the scheduler queue, stopping that task from being scheduled.

This solution is a band-aid, and more debugging should be done to determine why the task
was marked as queued in the first place.

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>