Remove a task from the task list by using LinkedList::remove
Before this was done by splitting the list into two, popping the first
element of the second list, then joining the lists back together.
This was cumbersome and verbose, so a smart programmer added a direct
remove function to the LinkedList struct.
Thanks to that programmer, we can use LinkedList::remove in our code to
make deleting a task cleaner.

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>