Make syscalls run in a normal context instead of in the interupt context
This gives us a couple of nice properties:
1) You generally want to have interrupts return quickly.
Now we don't have to worry about about how long syscalls take
2) You can now call syscalls in syscalls!
3) Some atomic data structures (e.g. crossebam queues) block a thread until
there is no contention on that structure. This could be problematic if this happens
in an inturrupt because the program would just hang.
Now, if a data structure has contention in the middle of a system call, that thread
will simply be rescheduled and whatever thread is using it can finish.

Signed-off-by: Bijan Tabatabai <bijan311@yahoo.com>