Fix race condition where a task would be mistakenly dropped
Sometimes, when scheduler() was called there would somehow only be one strong
reference to the task being held by NEXT_TASK. Because of this, when NEXT_TASK
was reassigned, the task it held previously would drop, freeing all the memory
associated with said task. However, this seems to be sporadic because this patch's
solution of not changing the running task when the strong reference count is one
works.
I do not have a good reason as to why there would only be one strong reference
sometimes, since the task list should always hold a reference to every active
task.
Maybe the underlying cause of this bug is a race condition in LinkedList? Or
maybe a bug in QEMU? I don't know.

Signed-off-by: Bijan Tabatabai <bijan311@gmail.com>