Free a task's dynamically allocated memory when it exits
There are a few problems I have with this solution, but I think it is passable for now.
The first problem I have is that the ALLOC_LIST is a Vec. It would be much more efficient
to use a binary tree, but for some reason when I tried using BTreeMap, FeRTOS would crash.
I'll investigate this for the next release.
Second, With this implementation, if two tasks share data via smart pointer types like
Arcs, the task that first created the smart pointer must live longer than the tasks
it shares the Arc with. If the task that created the Arc exits prematurely, we risk using
the data after it is freed.
Similarly, if a task spawns another task with a dynamic heap, the parent will have to
live longer than the child because the parent allocates the child's stack.

It's important to note that it is still possible for data to leak in these conditions,
but with this patch it is MUCH less likely to do so.

Signed-off-by: Bijan Tabatabai <bijan311@gmail.com>