Change the allocator to use a free list instead of in heap blocks.
This new heap works as follows:
    ---------------------------------------------
   | free list |         useable heap            |
    ---------------------------------------------
   The beginning of the heap is reserved for the free list.
   The ith bit of the free list is 0 if the ith block in the useable heap is
   unallocated and is 1 otherwise.
   The "useable heap" is the part of the heap that will be used to service
   dynamic memory allocations.

The main advantage of this is that if a naughty task were to accidentaly write
to the data outside its allocated region, it has a much lower chance of overwritting
heap metadata. A very convinient consequence to this is that FeRTOS will be more likely
to detect stack overflows because it won't be locked up in the allocator (though it
still can't catch every stack overflow).
Another advantage is that other than setting up the heap, I think the code is now simpler.

However, an obvious disadvantage is that allocations and deallocations will likely be slower,
and this strategy does not scale well with large heaps. However, FeRTOS is designed
for embedded applications, so that might not be that much of a problem.

Signed-off-by: Bijan Tabatabai <bijan311@gmail.com>